A0-2985: Refactor AZERO.ID components; Fix add address label text
