+++ packages/react-components/src/AzeroId/index.ts
@@ -1,0 +1,2 @@
+export { AzeroIdDomain, AzeroIdInteractiveDomain, AzeroIdRegisterLink } from './Atoms.js';
+export { default } from './AzeroId.js';
