+++ packages/react-components/src/index.ts
@@ -12,7 +12,7 @@ export { default as AddressSmall } from './AddressSmall.js';
 export { default as AddressToggle } from './AddressToggle.js';
 export { default as Available } from './Available.js';
 export { default as AvatarItem } from './AvatarItem.js';
-export { default as AzeroId, AzeroIdDomain } from './AzeroId.js';
+export { default as AzeroId, AzeroIdDomain } from './AzeroId/index.js';
 export { default as Badge } from './Badge.js';
 export { default as Balance } from './Balance.js';
 export { default as BatchWarning } from './BatchWarning.js';
